requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
openpyxl>=3.0.10
pdfplumber>=0.9.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pytest>=7.0.0
flake8>=4.0.0
